Refactored sun/moon logic into separate file
